refactor: 더 명확한 모듈 분리를 위해 폴더 레이어를 재구성
- 불필요한 repository, service, controller 폴더 제거
- User Domain에 종속된 validation을 Common으로 이동
