+++ src/auth/auth.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './service/auth.service';
+import { AuthService } from './auth.service';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigService } from '@nestjs/config';
-import { RefreshTokenRepository } from './repository/refresh-token.repository';
+import { RefreshTokenRepository } from './refresh-token.repository';
 import { PrismaModule } from '@app/prisma/prisma.module';
 
 @Module({
