+++ src/auth/auth.service.ts
@@ -7,7 +7,7 @@ import {
 } from '@nestjs/jwt';
 import * as bcrypt from 'bcrypt';
 import { randomUUID } from 'crypto';
-import { RefreshTokenRepository } from '../repository/refresh-token.repository';
+import { RefreshTokenRepository } from './refresh-token.repository';
 import { JwtPayload } from '@app/common/types/auth-user';
 import {
   RefreshTokenRevokedError,
