+++ src/common/validators/is-image-storage-url.constraint.ts
