+++ src/common/validators/is-image-storage-url.ts
