+++ src/user/__test__/mocks/auth-service.mock.ts
@@ -1,4 +1,4 @@
-import { AuthService } from '@app/auth/service/auth.service';
+import { AuthService } from '@app/auth/auth.service';
 
 export type MockAuthService = {
   [K in keyof AuthService]: jest.MockedFunction<AuthService[K]>;
